+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add GatheringICECandidate Finish Duration to Meeting Event and to demo app
+- Add attendeePresenceReceived, audioInputSelected, videoInputSelected, audioInputUnselected, videoInputUnselected
+- Compute and add totalStartDurationMs as part of the attributes of attendeePresenceReceived
 
 ### Changed